Trait packed::Unaligned

pub unsafe trait Unaligned { }

A marker trait indicating that a type has an alignment of 1.

In general, only applies to (), bool, i8, u8, and any types that contain only members of these types.