Crate packed

A safe approach to using #[repr(packed)] data.

See nue_macros for the automagic #[packed] attribute.

Traits

Aligned
A marker trait indicating that a type has an alignment over 1, and is therefore not safe to use in an unaligned context.

Packed
A marker trait indicating that a type is #[repr(packed)].

Unaligned
A marker trait indicating that a type has an alignment of 1.

Type Definitions

Un
A type alias that represents the unaligned type of T.